+++ PresentPicker/src/category/Sport.java
@@ -13,7 +13,7 @@ public class Sport extends Product {
 	
 	/* Full constructor, including those in the superclass Products */
 	public Sport (String equipmentFor, String sizeSport, String nameProduct, String descriptionProduct, 
-			String storeProduct, String imageProduct, float priceProduct) {
+			String storeProduct, String imageProduct, double priceProduct) {
 		super(nameProduct, descriptionProduct, storeProduct, imageProduct, priceProduct);
 		this.equipmentFor = equipmentFor;
 		this.sizeSport = sizeSport;
